kafka/requirements.txt
kafka-python
redis
confluent_kafka
python-dotenv
